+++ PPTX-to-PDF-conversion/Multithreading-using-parallel-process/.NET/Multithreading-using-parallel-process/Program.cs
@@ -5,26 +5,25 @@ using System.IO;
 using System.Threading.Tasks;
 using Syncfusion.PresentationRenderer;
 
-namespace Multithreaded_using_tasks
+namespace Multithreading_using_parallel_process
 {
     class MultiThreading
     {
-        //Indicates the number of threads to be create.
-        private const int TaskCount = 1000;
-        public static async Task Main()
+        static void Main(string[] args)
         {
-            //Create an array of tasks based on the TaskCount.
-            Task[] tasks = new Task[TaskCount];
-            for (int i = 0; i < TaskCount; i++)
+            //Indicates the number of threads to be create.
+            int limit = 5;
+            Console.WriteLine("Parallel For Loop");
+            Parallel.For(0, limit, count =>
             {
-                tasks[i] = Task.Run(() => ConvertPowerPointToPDF());
-            }
-            //Ensure all tasks complete by waiting on each task.
-            await Task.WhenAll(tasks);
+                Console.WriteLine("Task {0} started", count);
+                //Convert multiple presentations, one PPT on each thread.
+                ConvertPowerPointToPDF(count);
+                Console.WriteLine("Task {0} is done", count);
+            });
         }
-
         //Convert a Powerpoint presentation to PDF using multi-threading.
-        static void ConvertPowerPointToPDF()
+        static void ConvertPowerPointToPDF(int count)
         {
             using (FileStream inputStream = new FileStream(Path.GetFullPath(@"Data/Input.pptx"), FileMode.Open, FileAccess.Read))
             {
@@ -34,7 +33,7 @@ namespace Multithreaded_using_tasks
                     //Convert PowerPoint presentation to PDF.
                     using (PdfDocument pdfDocument = PresentationToPdfConverter.Convert(presentation))
                     {
-                        using (FileStream outputFileStream = new FileStream(Path.GetFullPath(@"Output/Output" + Guid.NewGuid().ToString() + ".pdf"), FileMode.Create, FileAccess.Write))
+                        using (FileStream outputFileStream = new FileStream(Path.GetFullPath(@"Output/Output" + count + ".pdf"), FileMode.Create, FileAccess.Write))
                         {
                             //Save the PDF document.
                             pdfDocument.Save(outputFileStream);
