+++ Read-and-save-PowerPoint-presentation/Multithreading-using-tasks/.NET/Multithreading-using-tasks/Multithreading-using-tasks.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net8.0</TargetFramework>
-    <RootNamespace>Multithreaded_using_tasks</RootNamespace>
+    <RootNamespace>Multithreading_using_tasks</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
